remove release with status decorator
